Small CFF cleanups
No real functional changes. Derives `Clone` for a few types, renames `Scaler` -> `Outlines` and moves some code between files.
